% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestDataBasalArea.R
\name{basalArea}
\alias{basalArea}
\title{Basal Area}
\usage{
basalArea(Measures, Surface)
}
\arguments{
\item{Measures}{Measures vector contain a list of circumferences}

\item{Surface}{Surface contain a surface}
}
\value{
Numeric Value
}
\description{
basalArea function calculate the basal area from 2 vectors of data
}
\examples{
basalArea(c(20,25,30,70,40,20,35),50)
15.99507
}
