% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{calc_dominant_height}
\alias{calc_dominant_height}
\title{Calculates dominant height}
\usage{
calc_dominant_height(nmax, ntress, height)
}
\value{
A numeric \code{vector}
}
\description{
Calculates dominant height
}
\keyword{internal}
