% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chorological_species}
\alias{chorological_species}
\title{Species list - Chorological Maps}
\format{
## `chorological_species`
A tibble with 104 rows and 3 columns:
\describe{
  \item{Species}{Latin name of tree species for \code{get_chorological}}
  \item{DOI}{Url to the source of the species' data}
}
}
\source{
<https://forest.jrc.ec.europa.eu/en/european-atlas/atlas-data-and-metadata/>
}
\usage{
chorological_species
}
\description{
A \code{tibble} with the names of the tree species that have a
chorological map available
}
\keyword{datasets}
