% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestDataClear.R
\name{clearCircumference}
\alias{clearCircumference}
\title{Clear Circumference}
\usage{
clearCircumference(data, idTree = "idTree", Measure = "Circ",
  MeasureYear = "CensusYear", Status = FALSE, idTaxon = FALSE,
  Genus = "Genus", Specie = "Species", returnFullDataframe = TRUE)
}
\arguments{
\item{data}{Dataframe containing data to clear}

\item{idTree}{Name of the id of the Tree column in the dataframe}

\item{Measure}{Name of the Measure column in the dataframe}

\item{MeasureYear}{Name of the Measure Date column in the dataframe}

\item{Status}{Name of the status of the tree column in the dataframe (if it doesn't exists in your dataset the function generate it from the NA in the dataframe)}

\item{idTaxon}{Name of the id of the taxon column in the dataframe (if it doesn't exists in your dataset you must specify the column of the genus and the specie)}

\item{Genus}{Name of the Genus column in the dataframe (if it doesn't exists in your dataset you must to specify the column name of idTaxon)}

\item{Specie}{Name of the Specie column in the dataframe (if it doesn't exists in your dataset you must to specify the column name of idTaxon)}

\item{replace}{When set to false it create a column circumf_corr, when set to true it replace measure named column with the corrected values}
}
\value{
a Dataframe cleared
}
\description{
Clear Circumference
}
\examples{
clearCircumference(data, idTree = "idTree", Measure = "Circ", MeasureYear = "Year", Status = "Status", Genus = "Genus", Specie = "Specie", replace = FALSE)

}
