% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ba.R
\name{compute_ba}
\alias{compute_ba}
\title{Title}
\usage{
compute_ba(data, measure = "Diameter", surface = NA,
  forest_col = "Forest", forest = NA, plot_col = "Plot", plot = NA,
  subplot_col = "Subplot", subplot = NA, time_col = "CensusYear",
  time = NA, taxon_col = c("Genus", "Species"), taxon = NA,
  Dmin = NA, Dmax = NA, categories = NA, aggregation = T)
}
\arguments{
\item{categories}{}
}
\description{
Title
}
