% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_growth.R
\name{compute_growth_diameter}
\alias{compute_growth_diameter}
\title{Title}
\usage{
compute_growth_diameter(data, id_col = "idTree",
  time_col = "CensusYear", size_col = "CircCorr", measure_type = "C")
}
\arguments{
\item{measure_type}{}
}
\description{
Title
}
