% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mortality.R
\name{compute_mortality}
\alias{compute_mortality}
\title{Title}
\usage{
compute_mortality(data, alive_col = "status_corr", time_col, id_col,
  dead_confirmation_censuses = 2, byplot = TRUE, plot_col = "plot",
  corrected = TRUE)
}
\arguments{
\item{corrected}{}
}
\description{
Title
}
