% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rates.R
\name{compute_rates}
\alias{compute_rates}
\title{Title}
\usage{
compute_rates(
  data,
  status_col = "status_corr",
  time_col = ifelse(is.null(getOption("time_col")), "CensusYear", getOption("time_col")),
  id_col = ifelse(is.null(getOption("id_col")), "idTree", getOption("id_col")),
  dead_confirmation_censuses = 2,
  byplot = TRUE,
  plot_col = ifelse(is.null(getOption("plot_col")), "Plot", getOption("plot_col")),
  corrected = TRUE
)
}
\arguments{
\item{data}{data.frame, containing forest inventories in the form of a
long-format time series - one line corresponds to a measurement for one
individual at a given census time.}

\item{status_col}{character, name of the column corresponding to tree status:
0/FALSE for dead, 1/TRUE for alive.}

\item{time_col}{character, name of the column containing census years.}

\item{id_col}{character, name of the column containing trees unique IDs.}

\item{dead_confirmation_censuses}{integer, defaults to 2: number of
consecutive censuses for which a tree is unseen that are needed to consider
the tree as dead. NB: for the trees unseen during the
dead_confirmation_censuses -1 last inventories, the status cannot be
corrected, thus mortality rates should not be calculated for these
censuses.}

\item{byplot}{logical, indicating whether the function has to process the
data by plot (TRUE)or for the whole dataset (FALSE).}

\item{plot_col}{character, name of the column containing plot indices or
names.}

\item{corrected}{Logical, indicates whether the dataset has been corrected
(for tree status errors) beforehand. If TRUE, triggers correct_alive,
defaults to TRUE.}
}
\value{
a data.frame that contains recruitment and mortality rates, in the
  same format as the outputs of compute_mortality and compute_recruitment
}
\description{
Title
}
\examples{
data(example_status_corr)
suppressWarnings(
compute_rates(example_status_corr))
}
