% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_recruitment.R
\name{compute_recruitment}
\alias{compute_recruitment}
\title{Title}
\usage{
compute_recruitment(data, alive_col = "status_corr", time_col, id_col,
  dead_confirmation_censuses = 2, byplot = TRUE, plot_col = "plot",
  corrected = TRUE)
}
\arguments{
\item{corrected}{}
}
\description{
Title
}
