% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_alive.R
\name{correct_alive}
\alias{correct_alive}
\title{Correct Trees Life Status in a Forest Inventory}
\usage{
correct_alive(data, id_col = "idTree", time_col = "CensusYear",
  alive_col = "CodeAlive", plot_col = "Plot", byplot = TRUE,
  dead_confirmation_censuses = 2, use_size = FALSE)
}
\arguments{
\item{data}{Data.frame, no default. Forest inventory in the form of a long-format time series - one line is a measure for one individual during one census time.}

\item{id_col}{Character. The name of the column containing trees unique ids}

\item{time_col}{Character. The name of the column containing census year}

\item{alive_col}{Character. The name of the column containing tree vital status - 0=dead; 1=alive.}

\item{plot_col}{Character. The name of the column containing the plots indices.}

\item{byplot}{Logical. If there are several plots in your dataset, the correction is performed by plot, in case these would not be censuses the same years or with the same frequencies one another.}

\item{dead_confirmation_censuses}{Integer, defaults to 2. This is the number of censuses needed to state that a tree is considered dead, if unseen. In Paracou, we use the rule-of-thumb that if a tree is unseen twice, its probability to be actually dead is close to 1. The choice of this value involves that trees unseen during the X-1 last inventories can not be corrected for death, and thus mortality rates should not be calculated for these censuses.}

\item{use_size}{Character, but defaults to FALSE. Optional argument specifying that circumference or diameter must be used to create the vital status field. If your data already contains a field indicating whether the tree is dead -0 or FALSE- or alive -1 or TRUE-, let it to its default value. If you use this option, make sure beforehand that only live trees are measured - non-NA size - in your dataset's protocol.}
}
\value{
a data.frame containing the corrected data - with trees' corrected life statuses. 1 = alive, 0 = dead. NAs indicate that the tree was unseen and cannot be considered yet. The output does not necessarily have the same number of lines as the input. Lines are added when the trees were unseen then seen alive again, with all columns being NA except trees' id, plot, census year and corrected status. Useless lines -with NA status before first sight alive, or after death- are suppressed.
}
\description{
Spotting unseen trees and correcting their status.
}
