% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_all.R
\name{correct_all}
\alias{correct_all}
\title{Full correction of a Forest Inventories Dataset}
\usage{
correct_all(data, id_col = "idTree", time_col = "CensusYear",
  status_col = "CodeAlive", POM_col, size_col = "Circ",
  measure_type = "circumference", plot_col = "Plot", byplot = TRUE,
  dead_confirmation_censuses = 2, positive_growth_threshold,
  negative_growth_threshold, default_POM, dbh_min = 10,
  use_size = FALSE)
}
\arguments{
\item{use_size}{}
}
\description{
Performs corrections for tree life status, overgrown recruits cases, size measurement errors and POM changes.
}
