% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_recruits.R
\name{correct_recruits}
\alias{correct_recruits}
\title{Correct Overgrown Recruits in Forest Inventories}
\usage{
correct_recruits(data, dbh_min = 10, diameter_growth_limit = 5,
  time_col = "CensusYear", id_col = "id", plot_col = "Plot",
  size_col = "Circ", status_col = "AliveCode",
  measure_type = "circumference", byplot = TRUE, corrected_status)
}
\arguments{
\item{data}{A data.frame containing a forest inventory in which each line corresponds to an individual's measurement for one census}

\item{dbh_min}{A scalar, integer or numeric, indicating the minimum DIAMETER at breast height at which trees are recorded, in centimeters. Defaults to 10 cm.}

\item{time_col}{A character indicating the name of the column containing census year information.}

\item{id_col}{A character indicating the name of the column containing individual tree's IDs}

\item{plot_col}{A character indicating the name of the column containing plot indices or names.}

\item{size_col}{A character indicating the name of the column containing tree size measurements -diameters or circumferences-.}

\item{status_col}{A character indicating the name of the column containing tree vital status - 0 for dead, 1 for alive, NA for unseen. No tree must appear with NA status before their first measurement.}

\item{measure_type}{A character indicating the size measurement type : "C" for circumference, "D" for diameter. Defaults to "C".}

\item{byplot}{A logical indicating whether there are several plots in the dataset. Defaults to TRUE}

\item{corrected_status}{A logical indicating whether the dataset was corrected for tree vital status errors beforehand. If FALSE, the correct_mortality function is called first with default parameters - see the function info section.}

\item{growth_limit}{A scalar, integer or numeric, indicating the maximum tolerated annual DIAMETER growth rate -in centimeters- over which a tree's growh is considered abnormal. Defaults to 5 cm, as set for Paracou inventories.}
}
\description{
Spot individuals that have first been unseen during previous censuses but were already over the minimum counting dbh, and provide a correction that is useful to compute biomass estimates.
}
