% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{crop_with_feedback}
\alias{crop_with_feedback}
\title{Crop a list of SpatRasters
(Internal) Crop using cli feedback}
\usage{
crop_with_feedback(r, xwgs84, quiet)
}
\value{
An \code{SpatRaster}
}
\description{
Crop a list of SpatRasters
(Internal) Crop using cli feedback
}
\keyword{internal}
