% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_corrected_sizes.R
\name{display_corrected_trees}
\alias{display_corrected_trees}
\title{Title}
\usage{
display_corrected_trees(
  data,
  code_corr_col = "code_corr",
  size_corr_col = "size_corr",
  size_col,
  plot_col = "plot",
  measure_type = "Circ",
  time_col = "CensusYear",
  status_corr_col = "status_corr",
  id_col = "id",
  info_cols = c("species", "plot"),
  path_save = "./graphs",
  name = "diameter_corrections_",
  tag_pom = FALSE,
  create_folder = TRUE,
  overwrite = TRUE,
  device = "jpeg",
  save_graph = TRUE
)
}
\arguments{
\item{data}{data.frame, containing forest inventories in the form of a
long-format time series - one line corresponds to a measurement for one
individual at a given census time.}

\item{code_corr_col}{Single character, the name of the column containing tree correction codes (see the output of correct_size)}

\item{size_corr_col}{Single character, the name of the column containing tree corrected size measurements}

\item{size_col}{character, name of the column corresponding to tree size
(circumference or diameter) measurements .}

\item{plot_col}{character, name of the column containing plot indices or
names.}

\item{measure_type}{character, partially matching “Circumference” or
“Diameter”, indicating what is the type of the measurements.}

\item{time_col}{character, name of the column containing census years.}

\item{status_corr_col}{Single character, the name of the column containing tree corrected status}

\item{id_col}{character, name of the column containing trees unique IDs.}

\item{info_cols}{Single or multiple character, the name(s) of the column containing the information that has to appear in the graphs' subtitle.}

\item{path_save}{Relevant if \code{save_graph=TRUE}. Character, a path
indicating in which FOLDER the graph has to be saved.}

\item{name}{Relevant if \code{save_graph=TRUE}. Character, the name of the
folder containing the graph. It can be followd by the extension
corresponding to the device - avoid .jpg for the jpeg device, use .jpeg
instead. If the extension is missing, it is automatically added according
to the selected device.}

\item{tag_pom}{Logical, whether POM shifts should be displayed with a vertical line on the graph, or not. Use only if you have a POm field in your dataset.}

\item{create_folder}{Relevant if \code{save_graph=TRUE}. Logical, indicated
whether the folders in the given path must be created in case they do not
exist yet, or not.}

\item{overwrite}{Relevant if \code{save_graph=TRUE}. Logical, indicating
whether a file already existing under the same name must be overwritten, or
kept. In the second case, the function aborts with an explicit error
message.}

\item{device}{Relevant if \code{save_graph=TRUE}. Character, the graphical
device to be used to save the graph.}

\item{save_graph}{Logical, indicates whether the graph must be saved or not.
Defaults to \code{FALSE}. If \code{TRUE}, please set the above described
arguments in an appropriate way.}
}
\value{
Nothing, in practical terms, this function is designed to save the graphs directly into a directory.
}
\description{
display_corrected_sizes allows to visualise the corrections performed on tree size measurements using ggplot2 utilities to make a points graph displaying raw (uncorrected) and corrected measurement series for censused trees, with all the specified relevant information.
}
