% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_growth.R
\name{display_growth}
\alias{display_growth}
\title{Display annual growth rates using ggplot2}
\usage{
display_growth(
  growth,
  type = "ribbon",
  time_col = "time",
  color_col = "Plot",
  growth_variable = "mean",
  faceting = FALSE,
  title = "Annual growth in function of census intervals.",
  subtitle = NULL,
  save_graph = FALSE,
  device = "png",
  path_save = file.path("ForestGraphs", paste0("annual_mortality_recruitment_", type,
    ".png")),
  name = "annual_growth.png",
  create_folder = FALSE,
  overwrite = FALSE,
  ordered_percentiles = c("P5", "P95"),
  ...
)
}
\arguments{
\item{growth}{A data.frame outputed from compute_growth}

\item{type}{Character, the type of graph (recommended: "line" or "ribbon")}

\item{time_col}{Character, name of the column corresponding to census time}

\item{color_col}{Character, name of the colomn used to define lines' colors,
defaults to "Plot".}

\item{growth_variable}{Character}

\item{faceting}{Character, name of the variable used for faceting -see after-
but defaults to FALSE i.e. no faceting. NB: faceting refers to using a
grouping variable to layout multiple plots, each corresponding to a
category of the grouping variable. See details for a pratical explanation.
The scales are free on the x axis: t correspond to several groups not
necessarily having the same censusing temporal resolution; but are bound on
the y-axis: the values are displayed on the same scale for comparison
purposes.}

\item{title}{Character,title of the graph.}

\item{subtitle}{Character, subtitle of the graph. Defaults to \code{NULL}}

\item{save_graph}{Logical, indicates whether the graph must be saved or not.
Defaults to \code{FALSE}. If \code{TRUE}, please set the above described
arguments in an appropriate way.}

\item{device}{Relevant if \code{save_graph=TRUE}. Character, the graphical
device to be used to save the graph.}

\item{path_save}{Relevant if \code{save_graph=TRUE}. Character, a path
indicating in which FOLDER the graph has to be saved.}

\item{name}{Relevant if \code{save_graph=TRUE}. Character, the name of the
folder containing the graph. It can be followd by the extension
corresponding to the device - avoid .jpg for the jpeg device, use .jpeg
instead. If the extension is missing, it is automatically added according
to the selected device.}

\item{create_folder}{Relevant if \code{save_graph=TRUE}. Logical, indicated
whether the folders in the given path must be created in case they do not
exist yet, or not.}

\item{overwrite}{Relevant if \code{save_graph=TRUE}. Logical, indicating
whether a file already existing under the same name must be overwritten, or
kept. In the second case, the function aborts with an explicit error
message.}

\item{ordered_percentiles}{Numeric, if plotting a ribbon, please indicate the percentiles used when aggregating the data.}

\item{...}{Additional arguments. Notably, verbose = F turns most warnings off}
}
\value{
A ggplot graph
}
\description{
Display annual growth rates using ggplot2
}
\examples{

data(example_size_corr)
growth <- suppressWarnings(compute_growth(example_size_corr,
                         size_col = "size_corr",
                         measure_type = "cir",
                         status_col = "CodeAlive",
                         id_col= "idTree",
                         time_col = "CensusYear",
                         what_output = "annual",
                         aggregate = TRUE,
                         by = c("Plot"),
                         stat = "mean",
                         percentiles = c(5,95)))
 display_growth(growth)
}
