% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_mortality.R
\name{display_mortality}
\alias{display_mortality}
\title{Display Mortality Rates with ggplot2}
\usage{
display_mortality(
  mortality = NULL,
  type = "line",
  time_col = "time",
  color_col = "plot",
  faceting = FALSE,
  title = "Annual mortality and recruitment rates in function of census intervals.",
  subtitle = NULL,
  save_graph = FALSE,
  device = "png",
  path_save = file.path("ForestGraphs", paste0("annual_mortality_recruitment_", type,
    ".png")),
  name = "Annual Mortality and Recruitment rates for each between-census interval",
  create_folder = FALSE,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{mortality}{data.frame, output of compute_mortality}

\item{type}{character, partially matching one of the following: "line", "histogram", "bar", "smooth". For the moment, only "line" is implemented, being the most relevant in this case. Corresponds to the type of graph to be done, for examples see ggplot2 examples for geom_line, geom_histogram, geom_bar and geom_smooth.}

\item{time_col}{Character, name of the column corresponding to census time}

\item{color_col}{Character, name of the colomn used to define lines' colors, defaults to "Plot".}

\item{faceting}{Character, name of the variable used for faceting -see after- but defaults to FALSE i.e. no faceting. Faceting refers, here, to using a categorical, grouping variable to layout multiple plots, each corresponding to a category of the grouping variable. For example, one can do a single graph per station -if there are several forest plots in each station-, or one per taxon if looking to several specific taxa' mortality and recruitment rates in detail. If offers the advantage to avoir plotting too much information on a graph, but it can lead to difficulties to compare lines with one another. The scales are free on the x axis -basically, it correspond to several groups not necessarily having the same censusing temporal resolution-, but are bound on y -i.e., the rates are displayed on the same scale for comparison purposes.}

\item{title}{Character,title of the graph.}

\item{subtitle}{Character, defaulting to null. The subtitle of the graph}

\item{save_graph}{Logical, indicates whether the graph must be saved or not. If TRUE, please set the above described arguments in an appropriate way.}

\item{device}{Character, the graphical device to be used to save the graph}

\item{path_save}{Character, a path indicating in which FOLDER the graph has to be saved}

\item{name}{Character, the name of the folder containing the graph. It can be followd by the extension corresponding to the device - avoid .jpg for the jpeg device, use .jpeg instead. If the extension is missing, it is automatically added according to the selected device.}

\item{create_folder}{Logical, indicated whether the folders in the given path must be created in case they do not exist yet, or not}

\item{overwrite}{Logical, indicating whether a file already existing under the same name must be overwritten, or kept. In the second case, the function aborts with an explicit error message.}
}
\value{
A ggplot2 graphical object.
}
\description{
Display Mortality Rates with ggplot2
}
