% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_recruits.R
\name{.corect_recruits_plot}
\alias{.corect_recruits_plot}
\title{Internal function. plot-level correction for overgrown recruits.}
\usage{
.corect_recruits_plot(data_plot, dbh_min, growth_limit)
}
\arguments{
\item{data_plot}{A data.frame containing a subset of forest inventory corresponding to one plot, formatted as in "correct_recruits".}

\item{dbh_min}{A scalar, integer or numeric, indicating the minimum DIAMETER at breast height at which trees are recorded, in centimeters.}

\item{growth_limit}{A scalar, integer or numeric, indicating the maximum tolerated annual DIAMETER growth rate -in centimeters- over which a tree's growh is considered abnormal.}
}
\value{

}
\description{
Internal function. plot-level correction for overgrown recruits.
}
