% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_alive.R
\name{.use_size_status}
\alias{.use_size_status}
\title{Create vital status field from size measurements under the hypothesis that only live trees were measured.}
\usage{
.use_size_status(data_plot, use_size)
}
\arguments{
\item{data_plot}{data.frame, a single-plot forest inventory whithout information on tree vital status, in which only live trees were measured, which names correspond to the format set in correct_alive, which is on the long format for census years - one line is one tree measured at one census.}

\item{use_size}{character, the name of the column containing tree size measurements - diameter of circumference.}
}
\value{
the same data.frame with a status field. status = 1 in any line where size is non-NA
}
\description{
Create vital status field from size measurements under the hypothesis that only live trees were measured.
}
