% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eutrees4f.R
\name{download_eutrees4f}
\alias{download_eutrees4f}
\title{(Internal) Downloads EU-Trees4F
 Download EU-Tree4F to tempdir}
\usage{
download_eutrees4f(download_url, dir_unzip, dir_zip, quiet)
}
\arguments{
\item{download_url}{Url of data to download}

\item{dir_unzip}{Path of the unzipped downloaded data}

\item{dir_zip}{Path of the zipped downloaded data}

\item{quiet}{If \code{TRUE} (the default), suppress status messages, and
the progress bar}
}
\value{
EU-Tree4F dir
}
\description{
(Internal) Downloads EU-Trees4F
 Download EU-Tree4F to tempdir
}
\examples{
\dontrun{
download_eutrees4f()
}
}
