% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_alive}
\alias{example_alive}
\title{Test cases for correct_alive}
\format{A data frame with 158 rows and 8 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
example_alive
}
\description{
A dataset containing examples to test correct_alive
}
\keyword{datasets}
