% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R, R/example_census-data.R
\docType{data}
\name{example_census}
\alias{example_census}
\title{Example dataset for ForestData: Paracou, plot 6}
\format{
A data frame with 25533 rows and 12 variables:
\describe{
  \item{Forest}{Character, general information: Forest sector where the plot
  to which trees belong is located. Here, always Paracou:P6}
  \item{Plot}{Integer (could be character), the index of the plot to which
  trees belong}
  \item{idTree}{Integer (could be character), the unique id of
  the tree}
  \item{Family}{Character, taxonomic information : botanical
  family}
  \item{Genus}{Character, taxonomic information : botanical genus}
  \item{Species}{Character, taxonomic information : species name}
  \item{CensusYear}{Integer, the year at which the tree is censused}
  \item{CodeAlive}{Logical or integer, tree life status: 1 for alive, 0 for
  dead. NA, if any, indicate that the tree was not seen for the corresponding
  census}
  \item{CodeMeas}{Character, not important here. Additionan
  informations relative to measurements}
  \item{Circ}{Numeric. Tree size
  expressed as stem circumference, in centimeters}
  \item{binomial_name}{Taxonomic information: full binomial species name}
  \item{POM}{Numeric. Point of Measurement, in meters.}
  }

An object of class \code{"data.frame"} with \itemize{
  \item{Forest}{The location of the plots included in the dataset}
  \item{Plot}{The plot indices correspondinc to the plot a given tree belongs
  to - In this dataset, the plots all correspond to subplots of Paracou's
  Plot6 (control, undisturbed)} \item{idTree}{Unique tree ID}
  \item{Family}{Taxonomical information, the plant family of each tree}
  \item{Genus}{Taxonomical information, the genus of each tree}
  \item{Species}{Taxonomical information, the species name of each
  tree} \item{binomial_name}{Taxonomical information, the complete binomial
  name (genus_species) of each tree} \item{CensusYear}{The census year at which each data have been acquired} \item{CodeAlive}{Whether the tree is dead (0) or alive (1). NAs do not exist in the original example dataset but can be found after using correct_alive (see the documentation)}
  \item{Circ}{The circumference of the tree} \item{POM}{The Point of
  Measurement (in meters) for each data}}
}
\source{
\url{https://paracou.cirad.fr/}

\href{https://paracou.cirad.fr/}{On demand}
}
\usage{
example_census

data(example_census)
}
\description{
A built-in example dataset used in ForestData's examples and vignette. It
corresponds to the squares 1 and 3 of the 6th Plot of the Paracou Experiment
database

This dataset was obtain from the Paracou database and exclusively aims at
illustrating ForestData features. It has been purposedly modified: plot 6
square 1 is provided as-is, but plot 6 square 3 misses several years to
simulate two plots with different censusing resolution.
}
\examples{
data(example_census)
}
\references{
The Paracou Research Station and Forest sDisturbance Experiment
  (\href{https://paracou.cirad.fr/}{Paracou Research Station})
}
\keyword{datasets}
