% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example_status_corr}
\alias{example_status_corr}
\title{Example dataset for ForestData: Paracou, plot 6, with corrected status}
\format{
A data frame with 25634 rows and 13 variables:
\describe{
  \item{Forest}{Character, Forest sector where the plot is located}
  \item{Plot}{Integer (could be character), the index of the plot to which
  trees belong}
  \item{idTree}{Integer (could be character), the unique id of
  the tree}
  \item{Family}{Character, taxonomic information : botanical
  family}
  \item{Genus}{Character, taxonomic information : botanical genus}
  \item{Species}{Character, taxonomic information : species name}
  \item{CensusYear}{Integer, the year at which the tree is censused}
  \item{CodeAlive}{Logical or integer, tree life status: 1 for alive, 0 for
  dead. NA, if any, indicate that the tree was not seen for the corresponding
  census}
  \item{CodeMeas}{Character, not important here. Additionan
  informations relative to measurements}
  \item{Circ}{Numeric. Tree size
  expressed as stem circumference, in centimeters}
  \item{binomial_name}{Taxonomic information: full binomial species name}
  \item{POM}{Numeric. Point of Measurement, in meters.}
  \item{status_corr}{Integer, interpretable as logical. Corrected vital
  status obtained with \code{correct_alive}}
  }
}
\source{
\url{https://paracou.cirad.fr/}
}
\usage{
example_status_corr
}
\description{
The same built-in example dataset as \code{example_census}, corrected with
\code{correct_alive}
}
\keyword{datasets}
