% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chorological_maps.R
\name{fd_forest_chorological}
\alias{fd_forest_chorological}
\title{Download the Chorological Maps}
\usage{
fd_forest_chorological(species, range = "nat")
}
\arguments{
\item{species}{The latin name of a tree species contained in the
Chorological Maps database (check \code{fd_forest_chorological_tbl()})}

\item{range}{The default "\code{nat}" downloads the probable native range
of the species, while "\code{syn}" downloads the synanthropic
range (i.e. the introduced and naturalized area and isolated
population since Neolithic)}
}
\value{
\code{sf} object
}
\description{
Download the Chorological Maps for the main European Woody Species
}
\examples{
\dontrun{
 # Download data for sweet chestnut
 chestnut_nat_sf <- fd_forest_chorological(species = "Castanea sativa", range = "nat")

 # Plot the data
 plot(chestnut_nat_sf$geometry)
}
}
