% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.R
\name{fd_forest_eu_biomass}
\alias{fd_forest_eu_biomass}
\title{European Maps of Forest Biomass}
\usage{
fd_forest_eu_biomass(
  x = NULL,
  lon = NULL,
  lat = NULL,
  layer = "AGB",
  year = 2023,
  crop = FALSE,
  mask = FALSE,
  merge = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{x}{a \code{sf} or \code{SpatVector} object. It will retrieve the
necessary tiles to cover the area (if \code{lat} and \code{lon} are
specified, this argument is ignored). See details.}

\item{lon}{a number specifying the longitude of the area where we want the tile}

\item{lat}{a number specifying the latitude of the area where we want the tile}

\item{layer}{a string for the layer to download. See details.}

\item{year}{a number specifying the year of the data. Possible options are 2017, and
2020-2023.}

\item{crop}{when \code{x} is specified, whether to crop the tile(s) to the object}

\item{mask}{when \code{x} is specified, whether to mask the tile(s) to the object}

\item{merge}{if \code{TRUE}, it will merge the tiles into one raster.
If \code{FALSE} a SpatRasterCollection will be returned.}

\item{quiet}{if \code{TRUE}, suppress any message or progress bar}
}
\value{
A \code{SpatRaster} or \code{SpatRasterCollection}
}
\description{
Download maps of Forest Biomass and Growing Stock Volume in Europe.
}
\details{
The data is distributed in tiles according to the Sentinel-2 tiles. Additionally, there's
a full map of Europe that can be accessed using the special option \code{x = "europe"}.

The available layers to download are:
\itemize{
\item \strong{AGB}: Aboveground Biomass
\item \strong{BGB}: Belowground Biomass
\item \strong{GSV}: Growing Stock Volume
}
}
\examples{
# download a tile in Galicia (Spain)
galicia_tile_rast <- fd_forest_eu_biomass(lon = -7, lat = 42)
}
\references{
\url{https://www.forestcarbonplatform.org/}
}
