% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eutrees4f.R
\name{fd_forest_eutrees4f}
\alias{fd_forest_eutrees4f}
\title{Download data from EU-Trees4F Database}
\usage{
fd_forest_eutrees4f(
  species,
  model = "clim",
  period = "all",
  scenario = "rcp45",
  type = "bin",
  distrib = "pot",
  quiet = TRUE
)
}
\arguments{
\item{species}{A character vector of length 1 with the latin name of the
tree species (genus and species)}

\item{model}{A character vector of length 1 with the name of the ensemble
projection. One of 'clim' or 'sdms' (see details)}

\item{period}{A numeric or character vector of length 1 with the center of
the 30-year time period used for the model. One of '2005',
'2035', '2065', '2095', or 'all' (see details)}

\item{scenario}{A character vector of length 1 with the climate change
scenario used. One of 'rcp45' or 'rcp85' (see details)}

\item{type}{A character vector of length 1 with the type of output layer.
One of 'bin', 'prob' or 'std' (see details)}

\item{distrib}{A character vector of length 1 with the type of distribution.
One of 'nat', 'pot', 'disp' or 'disp_lu' (see details)}

\item{quiet}{If \code{TRUE} (the default), suppress status messages, and
the progress bar}
}
\value{
A single-band or multi-band \code{SpatRaster}
}
\description{
Download data for tree species distribution in Europe for current (2005)
distribution, and future distribution (2035, 2065, 2095).
}
\details{
Data may be freely used for research, study, or teaching, but be cited
appropiately (see references below).


The data of EU-Trees4F database represent the distribution of the main woody
species in Europe at 5 arc-minutes (~ 10 km) spatial resolution, in the
Lambert Azimuthal Equal Area (EPSG:3035) CRS. The possible models to download
are the following:

__Model__: type of model used

- _clim_: climatic ensemble. A ensemble mean model that projects a consensus model
from \code{biomod2} into future conditions using the average of 11 Regional
Climate Models (RCM).

- _sdms_: Species Distribution Model (SDM) ensemble. A model that projects the
consensus model for every single RCM, and then it averages the output of
then 11 SDMs.

__Period__: 30-year time period

- _2005_: for current projections. This option ignores the scenario argument.
They are not available for \code{model = 'sdms'} with \code{type = 'std'}.

- _2035_: average of 2020-2050

- _2065_: average of 2050-2080

- _2095_: average of 2080-2110

- _all_: get the four periods (or three for std type)

__Scenario__: climate change scenario

- _rcp45_: a climate change scenario that assumes moderate emissions reductions

- _rcp85_: a climate change scenario with high greenhouse gas emissions
and limited mitigation efforts

__Type__: type of output layer

_bin_: binary distribution map, where 1 represents presence of the tree species,
while 0 represents absence of the tree species, derived from the prob map

_prob_: probability disitribution map (0-1000). Represents the probability
of being the potential distribution of the species

_std_: standard deviation of prob map. Only available for \code{model = 'sdms'}.

__Distrib__: type of species distribution

- _nat_: realized distribution (masked with native range). Only available
with \code{type = 'bin'}

- _pot_: potential distribution

- _disp_: natural dispersal model (migclim). Only available
with \code{type = 'bin'}

- *disp_lu*: natural dispersal model clipped by forest areas. Only available
with \code{type = 'bin'}
}
\examples{
\dontrun{
# Download data for Betula pendula
betula_pendula_sr <- fd_forest_eutrees4f(species = "Betula pendula")

}
}
\references{
Mauri, Achille; Cescatti, Alessandro; GIRARDELLO, MARCO; Strona,
Giovanni; Beck, Pieter; Caudullo, Giovanni; et al. (2022). EU-Trees4F. A
dataset on the future distribution of European tree species.. figshare.
Collection. https://doi.org/10.6084/m9.figshare.c.5525688.v2
}
