% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_land_cover.R
\name{fd_landcover_copernicus}
\alias{fd_landcover_copernicus}
\title{Download data form the Global Land Cover}
\usage{
fd_landcover_copernicus(
  x,
  lon = NULL,
  lat = NULL,
  year = 2019,
  layer = "forest",
  crop = FALSE,
  ...
)
}
\arguments{
\item{x}{An \code{sf} or \code{SpatVector} object. It will retrieve the
necessary tiles to cover the area (if \code{lat} and \code{lon} are
specified, this argument is ignored)}

\item{lon}{A number specifying the longitude of the area where we want the tile}

\item{lat}{A number specifying the latitude of the area where we want the tile}

\item{year}{Year of the forest extent data. One of 2015:2019 or 'all'}

\item{layer}{The layer to use from the Global Land Cover. See details}

\item{crop}{When \code{x} is specified, whether to crop the tiles(s) to the
object}

\item{...}{additional arguments passed to the \code{terra::crop} function}
}
\value{
\code{SpatRaster} object
}
\description{
Download a SpatRaster from the Global Land Cover from the Copernicus Global
Land Service.
}
\details{
The ....
}
\examples{
\dontrun{
 # Get tile for Galicia (Spain)
 galicia_forest_extent <- fd_landcover_copernicus(lat = 42.7, lon = -7.8, year = 2020)
 # Get masked tile for Galicia (Spain)
 galicia_forest_extent <- fd_landcover_copernicus(
 lat  = 42.7,
 lon  = -7.8,
 year = 2020,
 crop = TRUE,
 mask = TRUE)
}
}
\references{
Buchhorn, M.; Smets, B.; Bertels, L.; De Roo, B.; Lesiv, M.;
            Tsendbazar, N. - E.; Herold, M.; Fritz, S. Copernicus Global
            Land Service: Land Cover 100m: collection 3: epoch 2019: Globe
            2020. DOI 10.5281/zenodo.3939050
}
