% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{fdi_basal_area}
\alias{fdi_basal_area}
\title{Calculates Basal Area in square meters.}
\usage{
fdi_basal_area(diameter, ntrees = NULL, units = "cm")
}
\value{
A numeric \code{vector}
}
\description{
Calculates Basal Area in square meters.
}
\keyword{internal}
