% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{fdi_diametric_class}
\alias{fdi_diametric_class}
\title{Calculates diametric class}
\usage{
fdi_diametric_class(
  x,
  dmin = 7.5,
  dmax = NULL,
  class_length = 5,
  include_lowest = TRUE,
  return_intervals = FALSE
)
}
\value{
A numeric \code{vector}
}
\description{
Calculates diametric class
}
\keyword{internal}
