% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{fdi_dominant_height}
\alias{fdi_dominant_height}
\title{Calculates the dominant height}
\usage{
fdi_dominant_height(diameter, height, ntrees = NULL, which = "assman")
}
\value{
A numeric \code{vector}
}
\description{
Calculates the dominant height
}
\keyword{internal}
