% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{fdi_download}
\alias{fdi_download}
\title{(Internal) Downloads data to tempdir
Download data to tempdir}
\usage{
fdi_download(download_url, destfile, timeout = 1e+05)
}
\arguments{
\item{download_url}{Url of data to download}

\item{destfile}{Path of the downloaded data}

\item{timeout}{Time to stop downloading}
}
\value{
TRUE or FALSE
}
\description{
(Internal) Downloads data to tempdir
Download data to tempdir
}
\keyword{internal}
