% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_notExported.R
\name{fdi_download_7zip}
\alias{fdi_download_7zip}
\title{(Internal) Downloads data to tempdir
 Download data to tempdir}
\usage{
fdi_download_7zip(
  download_url,
  dir_unzip,
  dir_zip,
  timeout = 5000,
  quiet = TRUE
)
}
\arguments{
\item{download_url}{Url of data to download}

\item{dir_unzip}{Path of the unzipped downloaded data}

\item{dir_zip}{Path of the zipped downloaded data}

\item{timeout}{Time to stop downloading}

\item{quiet}{If \code{TRUE} (the default), suppress status messages, and
the progress bar}
}
\value{
Unzipped file
}
\description{
(Internal) Downloads data to tempdir
 Download data to tempdir
}
\keyword{internal}
