% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{fdi_download_raster}
\alias{fdi_download_raster}
\title{Donwload a read a raster
(Internal) Helper to download and read a raster from an URL}
\usage{
fdi_download_raster(url, start = NULL, end = NULL, timeout = 5000)
}
\value{
A \code{SpatRaster}
}
\description{
Donwload a read a raster
(Internal) Helper to download and read a raster from an URL
}
\keyword{internal}
