% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_notExported.R
\name{fdi_fix_names}
\alias{fdi_fix_names}
\title{(Internal) Fix non Latin-ASCII names}
\usage{
fdi_fix_names(name)
}
\arguments{
\item{name}{String to fix}
}
\value{
A character vector of same length as name
}
\description{
A function that fixes names with strange characters, spaces, and
also convert to title
}
\keyword{internal}
