% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestdata-package.R
\docType{package}
\name{forestdata-package}
\alias{forestdata}
\alias{forestdata-package}
\title{forestdata: Download Forestry Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions for downloading forestry and land use data for use in spatial analysis. This packages offers a user-friendly solution to quickly obtain datasets such as forest height, forest types, tree species under various climate change scenarios, or land use data among others.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cidree.github.io/forestdata/}
}

}
\author{
\strong{Maintainer}: Adrián Cidre González \email{adrian.cidre@gmail.com}

}
\keyword{internal}
