% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_country.R
\name{get_bdforet_tbl}
\alias{get_bdforet_tbl}
\title{(Internal) Creates the BD Forêt table}
\usage{
get_bdforet_tbl()
}
\value{
A \code{tibble}
}
\description{
Creates a table with the France department names, the url to download
the data, and the version
}
\examples{
\dontrun{
get_bdforet_tbl()
}
}
\keyword{internal}
