% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_notExported.R
\name{get_chorological_tbl}
\alias{get_chorological_tbl}
\title{Chorological Maps table
 Get a table with the names and references of each species}
\usage{
get_chorological_tbl()
}
\value{
\code{tibble}
}
\description{
Chorological Maps table
 Get a table with the names and references of each species
}
\examples{
\dontrun{
get_chorological_tbl()
}
}
