% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{get_combined_raster_2l}
\alias{get_combined_raster_2l}
\title{Combines different raster tiles
(Internal) Helper to combine rasters from Copernicus Global Land Cover.}
\usage{
get_combined_raster_2l(year_i, layer_i, url_table)
}
\value{
A \code{SpatRaster}
}
\description{
Combines different raster tiles
(Internal) Helper to combine rasters from Copernicus Global Land Cover.
}
\keyword{internal}
