% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eutrees4f.R
\name{get_eutrees4f_tbl}
\alias{get_eutrees4f_tbl}
\title{(Internal) Gets the tree species
 Get tree species for EU-Trees4F database}
\usage{
get_eutrees4f_tbl()
}
\value{
a character vector
}
\description{
(Internal) Gets the tree species
 Get tree species for EU-Trees4F database
}
\keyword{internal}
