% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLAD.R
\name{get_forest_extent_tbl}
\alias{get_forest_extent_tbl}
\title{Forest Extent Table
Get a table with the year, coordinates, and URL of the Forest Extent files
from the Global Land Analysis & Discovery (GLAD)}
\usage{
get_forest_extent_tbl()
}
\value{
A \code{tibble}
}
\description{
Forest Extent Table
Get a table with the year, coordinates, and URL of the Forest Extent files
from the Global Land Analysis & Discovery (GLAD)
}
\keyword{internal}
