% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy-height.R
\name{get_gch_tbl}
\alias{get_gch_tbl}
\title{Global Canopy Height Table
 Get a table with the coordinates and URL of the Global Canopy Height tiles}
\usage{
get_gch_tbl()
}
\value{
A \code{tibble}
}
\description{
Global Canopy Height Table
 Get a table with the coordinates and URL of the Global Canopy Height tiles
}
\references{
<https://gee-community-catalog.org/projects/canopy/>
<https://www.research-collection.ethz.ch/handle/20.500.11850/609802>
}
\keyword{internal}
