% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLAD.R
\name{get_glad_tiles}
\alias{get_glad_tiles}
\title{Global Land Analysis & Discovery tiles
Get an sf object with the tiles of GLAD datasets}
\usage{
get_glad_tiles()
}
\value{
An \code{sf} object
}
\description{
Global Land Analysis & Discovery tiles
Get an sf object with the tiles of GLAD datasets
}
\keyword{internal}
