% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land-cover.R
\name{get_glc_tbl}
\alias{get_glc_tbl}
\title{Global Land Cover Table
Get a table with the year, coordinates, and URL of the Global Land Cover
tiles.}
\usage{
get_glc_tbl()
}
\value{
A \code{tibble}
}
\description{
Global Land Cover Table
Get a table with the year, coordinates, and URL of the Global Land Cover
tiles.
}
\references{
\url{https://lcviewer.vito.be/download}
}
\keyword{internal}
