% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land-cover.R
\name{get_landcoverexplorer_tbl}
\alias{get_landcoverexplorer_tbl}
\title{(Internal) Get codes table of Land Cover Explorer}
\usage{
get_landcoverexplorer_tbl()
}
\value{
A \code{tibble}
}
\description{
Get a table with the Land Cover explorer codes and urls
}
\keyword{internal}
