% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy-height.R
\name{get_meta_tiles}
\alias{get_meta_tiles}
\title{Meta Global Canopy Height tiles
Get a vectorial object with the tiles of the dataset}
\usage{
get_meta_tiles()
}
\value{
An \code{sf} object
}
\description{
Meta Global Canopy Height tiles
Get a vectorial object with the tiles of the dataset
}
\references{
\url{https://registry.opendata.aws/dataforgood-fb-forests/}
}
\keyword{internal}
