% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_forest.R
\name{get_mfe50_ccaa_tbl}
\alias{get_mfe50_ccaa_tbl}
\title{(Internal) Get table CCAA/MFE50}
\usage{
get_mfe50_ccaa_tbl()
}
\value{
A \code{tibble}
}
\description{
Get a table with the Autonomous Communities of Spain, and the urls
to each of them.
}
\examples{
\dontrun{
get_mfe50_ccaa_tbl()
}
}
\keyword{internal}
