% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_forest.R
\name{get_mfe50_provinces_tbl}
\alias{get_mfe50_provinces_tbl}
\title{(Internal) Creates the MFE50 provinces table}
\usage{
get_mfe50_provinces_tbl()
}
\value{
A \code{tibble}
}
\description{
Creates a table with the province names, and the url to each province's
vectorial data (full table with all provinces)
}
\examples{
\dontrun{
get_mfe50_provinces_tbl()
}
}
\keyword{internal}
