% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_inventory.R
\name{get_spain_ifn2_tbl}
\alias{get_spain_ifn2_tbl}
\title{(Internal) Get table with Provinces and Urls for the Spanish IFN2}
\usage{
get_spain_ifn2_tbl()
}
\value{
A \code{tibble}
}
\description{
Get a table with the Provinces of Spain, and the Urls to download data
}
\keyword{internal}
