% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_inventory.R
\name{get_spain_ifn4_tbl}
\alias{get_spain_ifn4_tbl}
\title{(Internal) Get table Provinces and Urls}
\usage{
get_spain_ifn4_tbl()
}
\value{
A \code{tibble}
}
\description{
Get a table with the Provinces of Spain, and the Urls to download data
}
\keyword{internal}
