% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_inventory.R
\name{get_spain_ifn_metadata_tbl}
\alias{get_spain_ifn_metadata_tbl}
\title{(Internal) Get table with URLs of the IFN metadata}
\usage{
get_spain_ifn_metadata_tbl()
}
\value{
A \code{tibble}
}
\description{
Get a table with the IFN and database codes, and their metadata download URL
}
\keyword{internal}
