% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestDataBasalArea.R
\name{guyafor_BasalArea}
\alias{guyafor_BasalArea}
\title{Guyafor Basal Area}
\usage{
guyafor_BasalArea(data, forest, plot = F, subplot = F, taxon = F,
  year, surface, Dmin = F, Dmax = F)
}
\arguments{
\item{Dmax}{}
}
\value{

}
\description{
Guyafor Basal Area
}
\examples{
forestData.guyafor_BasalArea(data = DataGuyafor, forest = "Paracou", plot=1, subplot =  1, year =  2016, surface = 500, Dmin = 20)
}
