% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{mask_with_feedback}
\alias{mask_with_feedback}
\title{Masks a list of SpatRasters
(Internal) Masks using cli feedback}
\usage{
mask_with_feedback(r, xwgs84, quiet)
}
\value{
An \code{SpatRaster}
}
\description{
Masks a list of SpatRasters
(Internal) Masks using cli feedback
}
\keyword{internal}
