% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_forestdata}
\alias{metadata_forestdata}
\title{Metadata for `forestdata` functions}
\format{
## `metadata_forestdata`
A list of 7 elements:
\describe{
  \item{chorological_species}{Latin name of tree species for \code{fd_forest_chorological}}
  \item{eutrees4f_species}{Latin name of tree species for \code{fd_forest_eutrees4f}}
  \item{bdforet_tbl_departments}{Departments of France for \code{fd_forest_france}}
  \item{mfe_provinces}{Province names for \code{fd_forest_spain_mfe50}}
  \item{spain_ifnx}{Province names for \code{fd_inventory_spain}}
}
}
\usage{
metadata_forestdata
}
\description{
A \code{list} with the names of tree species or regions depending on the
dataset.
}
\keyword{datasets}
