% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{nest_ifn_regeneration}
\alias{nest_ifn_regeneration}
\title{Nest IFN3 or IFN 4 regeneration data
(Internal) Helper to nest data from IFN}
\usage{
nest_ifn_regeneration(data, codes, process_level = 1)
}
\value{
A \code{tibble}
}
\description{
Nest IFN3 or IFN 4 regeneration data
(Internal) Helper to nest data from IFN
}
\keyword{internal}
