% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{nest_ifn_shrub}
\alias{nest_ifn_shrub}
\title{Nest IFN3 or IFN 4 shrub data
(Internal) Helper to nest data from IFN}
\usage{
nest_ifn_shrub(data, codes, agents, ifn = 4)
}
\value{
A \code{tibble}
}
\description{
Nest IFN3 or IFN 4 shrub data
(Internal) Helper to nest data from IFN
}
\keyword{internal}
