% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{nest_ifn_tree}
\alias{nest_ifn_tree}
\title{Nest IFN3 or IFN 4 tree data
(Internal) Helper to nest data from IFN}
\usage{
nest_ifn_tree(
  data,
  codes,
  agents,
  which = "current",
  ifn = 4,
  process_level = 1
)
}
\value{
A \code{tibble}
}
\description{
Nest IFN3 or IFN 4 tree data
(Internal) Helper to nest data from IFN
}
\keyword{internal}
