% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestDataMap.R
\name{plotMap}
\alias{plotMap}
\title{plotMap}
\usage{
plotMap(data, X = "X", Y = "Y", IdTree = "idTree", Measures = F,
  Status = F, title = "Map", label_size = 2)
}
\arguments{
\item{data}{Dataframe containing the data of the plot}

\item{X}{Name of the X coordinate column in the dataframe}

\item{Y}{Name of the Y coordinate column in the dataframe}

\item{IdTree}{Name of the ID tree column in the dataframe}

\item{Measures}{Name of the measure column in the dataframe}

\item{Status}{Name of the status column in the dataframe}

\item{title}{Title of the plot, will be placed on the top of the draw}

\item{label_size}{Size of the label of each tree}
}
\value{
A ggplot2 plot
}
\description{
This function take a dataframe and draw a logging map
}
