% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{process_ifn}
\alias{process_ifn}
\title{Process a list returned by fd_iventory_spain
(Internal) Helper process IFN3 or IFN4 data}
\usage{
process_ifn(data, process_level = 1, ifn = 4, province_fix)
}
\value{
An \code{sf} object
}
\description{
Process a list returned by fd_iventory_spain
(Internal) Helper process IFN3 or IFN4 data
}
\keyword{internal}
