% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{process_pmayores}
\alias{process_pmayores}
\title{Level 2 process of tree data
(Internal) Helper to process tree data of IFN}
\usage{
process_pmayores(data)
}
\value{
A \code{tibble}
}
\description{
Level 2 process of tree data
(Internal) Helper to process tree data of IFN
}
\keyword{internal}
