% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rename_variable_col.R
\name{check_rename_variable_col}
\alias{check_rename_variable_col}
\title{Title}
\usage{
check_rename_variable_col(arg, argname, data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Title
}
