% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ba.R
\name{compute_ba}
\alias{compute_ba}
\title{Compute Basal Area in Forest Inventories}
\usage{
compute_ba(data, measure_col = "CircCorr", measure_type = "C",
  time_col = "CensusYear", by = c("Plot", "Subplot"), surface = F)
}
\arguments{
\item{data}{A data.frame containing a time-series tree-wise forest inventory -i.e. every line is a single tree measurement for a single year.}

\item{measure_col}{A single character containing the name of the column corresponding to tree size measurements -either circumference or diameter.}

\item{measure_type}{A single character indicating whether tree sizes are given in circumferences -"C"- or diameter -"D"-.}

\item{by}{A character vector containing the name of the columns containing the variables -other than census time- according to which the result will be aggregated. Be it plots, subplots or species name...}

\item{surface}{Either a scalar containing the surface area of each plot -if they have the same dimensions- or a data.frame of the surface area according to some of the grouping variables -e.g. Plot and subplot. However, defaults to FALSE and in this case, only absolute BA is returned.}
}
\value{
A dataframe containing, for each combination of grouping variables, the plot-level and per-ha basal area for a given census year.
}
\description{
Compute Basal Area according to user-specified grouping variables, and time.
}
\examples{
\dontrun{
data("Paracou6")
function(Paracou6,
measure_col = "CircCorr",
measure_type = "C",
time_col = "CensusYear",
by = c("Species"),
surface = 6.25)
}
}
