% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_recruits.R
\name{correct_recruits}
\alias{correct_recruits}
\title{Correct Overgrown Recruits in Forest Inventories}
\usage{
correct_recruits(data, dbh_min = 10, positive_growth_threshold = 5,
  time_col = "CensusYear", id_col = "idTree", plot_col = "Plot",
  size_corr_col = "size_corr", status_corr_col = "status_corr",
  measure_type = "circumference", invariant_columns = c("SubPlot",
  "species", "Genus", "Species"), byplot = TRUE,
  correct_status = FALSE)
}
\arguments{
\item{data}{data.frame, containing forest inventories in the form of a
long-format time series - one line corresponds to a measurement for one
individual at a given census time.}

\item{dbh_min}{scalar integer or numeric, indicating the minimum DIAMETER (in
centimeters) at the default measurement height from which trees are
recorded. Defaults to 10 cm.}

\item{positive_growth_threshold}{positive numeric or integer, threshold over
which an annual DIAMETER growth is considered abnormal (in cm). Defaults to
5 cm.}

\item{time_col}{character, name of the column containing census years.}

\item{id_col}{character, name of the column containing trees unique IDs.}

\item{plot_col}{character, name of the column containing plot indices or
names.}

\item{measure_type}{character, partially matching “Circumference” or
“Diameter”, indicating what is the type of the measurements.}

\item{invariant_columns}{character vector, containing the name of the columns
for which value remain constant for a given tree (for example species name
or coordinates). When a row is added by the function correct_alive, values
for invariant columns are taken from the value for other censuses. Defaults
to null}

\item{byplot}{logical, indicating whether the function has to process the
data by plot (TRUE)or for the whole dataset (FALSE).}
}
\value{
A corrected data.frame
}
\description{
correct recruits spots individuals that have first been unseen during previous censuses but were already over the minimum counting dbh, and provides a correction that is useful to compute biomass estimates.
}
\examples{
\dontrun{
data("Paracou6")
correct_recruits(Paracou6,
dbh_min = 10,
positive_growth_threshold = 5,
time_col = "CensusYear",
id_col = "id",
plot_col = "Plot",
size_corr_col = "Circ",
status_corr_col = "AliveCode",
measure_type = "circumference",
byplot = TRUE,
correct_status = FALSE)
}
}
