% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_size_main.R
\name{correct_size}
\alias{correct_size}
\title{Title}
\usage{
correct_size(data, size_col, time_col, status_col, id_col, POM_col,
  positive_growth_threshold, negative_growth_threshold, default_POM)
}
\arguments{
\item{default_POM}{}
}
\description{
Title
}
