% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_recruits.R
\name{.correct_recruits_tree}
\alias{.correct_recruits_tree}
\title{Internal function. tree-level correction for overgrown recruits.}
\usage{
.correct_recruits_tree(tree, dbh_min, growth_limit, censuses)
}
\arguments{
\item{tree}{A data.frame containing one single tree's time series of measurements.}

\item{dbh_min}{A scalar, integer or numeric, indicating the minimum DIAMETER at breast height at which trees are recorded, in centimeters.}

\item{growth_limit}{A scalar, integer or numeric, indicating the maximum tolerated annual DIAMETER growth rate -in centimeters- over which a tree's growh is considered abnormal.}

\item{censuses}{A numeric vector containing all the years for which the plot to which the tree belongs were censused.}
}
\value{
A data.frame containing the corrected indiv
}
\description{
Internal function. tree-level correction for overgrown recruits.
}
\examples{
#' \dontrun{
data("Paracou6")
cens <- unique(Paracou6$CensusYear)
ids <- unique(Paracou6$idTree)
for(i in ids) correct_recruits_tree(Paracou[Paracou$idTree == i,], 10, 5,cens)
}
}
