% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy_height.R
\name{fd_canopy_height}
\alias{fd_canopy_height}
\title{Canopy Height
 Get canopy height}
\usage{
fd_canopy_height(x, lon = NULL, lat = NULL, layer = "chm", crop = FALSE, ...)
}
\value{
A \code{SpatRaster}
}
\description{
Canopy Height
 Get canopy height
}
\examples{
\dontrun{
fd_canopy_height(lon = -7.27, lat = 42.43)
}
}
