% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_cover_explorer.R
\name{fd_landcover_esri}
\alias{fd_landcover_esri}
\title{Download a Land Cover SpatRaster for a UTM tile}
\usage{
fd_landcover_esri(utm_code, year, quiet = TRUE)
}
\arguments{
\item{utm_code}{A character string of length 1 with an UTM code (e.g. "29N)}

\item{year}{An integer or vector of integers corresponding to the base year
of the Land Cover image. The option \code{year = 'all'} downloads all
the available images}

\item{quiet}{If \code{TRUE} (the default), suppress status messages, and
the progress bar}
}
\value{
A \code{SpatRaster}
}
\description{
Download a UTM tile of the ESRI Land Cover Explorer for a specified year
}
\examples{
\dontrun{
# Download Land Cover for UTM tile 29N year 2022
lc <- fd_landcover_esri("29N", 2017:2020)
}
}
\references{
\url{https://livingatlas.arcgis.com/en/home/}
}
