% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not-exported.R
\name{get_combined_raster}
\alias{get_combined_raster}
\title{Combines different raster tiles
 (Internal) Helper to combine rasters from forest extent GLAD}
\usage{
get_combined_raster(year_i, url_table)
}
\value{
A \code{SpatRaster}
}
\description{
Combines different raster tiles
 (Internal) Helper to combine rasters from forest extent GLAD
}
\keyword{internal}
