% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_notExported.R
\name{get_combined_raster}
\alias{get_combined_raster}
\title{Combines different raster tiles
 (Internal) Helper to combine rasters from forest extent GLAD and
 Copernicus Global Land Cover.}
\usage{
get_combined_raster(year_i, url_table)
}
\value{
A \code{SpatRaster}
}
\description{
Combines different raster tiles
 (Internal) Helper to combine rasters from forest extent GLAD and
 Copernicus Global Land Cover.
}
\examples{
\dontrun{
get_combined_raster(2000, url_table = urls)
}
}
\keyword{internal}
